requirements.txt
flask
flask-sqlalchemy
flask-smorest
python-dotenv
marshmallow
pymysql
